+++ users/router.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from fastapi import APIRouter, HTTPException, Depends
+from fastapi import APIRouter, HTTPException, Depends, Query
 from fastapi.security import OAuth2PasswordBearer
 from starlette import status
 from starlette.exceptions import HTTPException
@@ -20,13 +20,13 @@ async def register(data: UserCreateSchema):
     if user:
         return user
     else:
-        raise HTTPException(status_code=400, detail="Failed to register user")
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Failed to register user")
 
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
 
-@users_router.post("/token")
+@users_router.post("/login")
 async def login_for_access_token(form_data: UserLoginSchema):
     user = await User.get(username=form_data.username)
     if user and pwd_context.verify(form_data.password, user.password):
@@ -36,18 +36,18 @@ async def login_for_access_token(form_data: UserLoginSchema):
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Incorrect username or password")
 
 
-@users_router.get("/protected-route")
-async def protected_route(token: str = Depends(oauth2_scheme)):
+@users_router.get('/me', response_model=UserListSchema)
+async def get_profile(token: str = Depends(oauth2_scheme)):
     payload = verify_token(token)
     if payload:
-        return {"message": "You are authorized"}
+        return await User.get(username=payload.get('sub'))
     else:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid token")
 
 
 @users_router.get('/', response_model=List[UserListSchema])
-async def get_users():
-    return await User.all()
+async def get_users(skip: int = Query(0, ge=0), limit: int = Query(10, ge=1)):
+    return await User.filter().offset(skip).limit(limit).all()
 
 
 @users_router.get('/{user_id}', response_model=UserListSchema)
@@ -57,15 +57,24 @@ async def get_user(user_id: int):
 
 
 @users_router.put('/{user_id}', response_model=UserListSchema)
-async def update_user(user_id: int, data: UserUpdateSchema):
-    user = await User.filter(id=user_id).update(**data.model_dump())
-    print(user)
-    return await User.get(id=user_id)
+async def update_user(user_id: int, data: UserUpdateSchema, token: str = Depends(oauth2_scheme)):
+    payload = verify_token(token)
+    user = await User.get(username=payload.get('sub'))
+    if user:
+        await User.filter(id=user_id).update(**data.model_dump())
+        return await User.get(id=user_id)
+    else:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='You have not enough permissions')
 
 
-@users_router.delete('/{user_id}', response_model=UserListSchema)
-async def delete_user(user_id: int):
-    deleted_count = await User.filter(id=user_id).delete()
-    if not deleted_count:
-        raise HTTPException(status_code=404, detail=f'Example {user_id} not found')
-    return Status(message=f'Example {user_id} deleted', details=None)
+@users_router.delete('/{user_id}', response_model=Status)
+async def delete_user(user_id: int, token: str = Depends(oauth2_scheme)):
+    payload = verify_token(token)
+    user = await User.get(username=payload.get('sub'))
+    if user.is_superuser:
+        deleted_count = await User.filter(id=user_id).delete()
+        if not deleted_count:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'User {user_id} not found')
+        return Status(message=f'User {user_id} deleted', details=None)
+    else:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='You have not enough permissions')
